Decompose stems, shift to util
