+++ redbiom/admin.py
@@ -356,44 +356,3 @@ def _stage_for_load(table, context, get, memberkey, tag=None):
 
     table.filter(to_load)
     return table.filter(lambda v, i, md: v.sum() > 0, axis='observation')
-
-
-def df_to_stems(df):
-    """Convert a DataFrame to stem -> index associations
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        A pandas DataFrame to index
-
-    Returns
-    -------
-    dict
-        {stem: {set of indices}}
-    """
-    import re
-    import nltk
-    from collections import defaultdict
-
-    # not using nltk default as we want this to be portable so that, for
-    # instance, a javascript library can query
-    p = nltk.PorterStemmer(nltk.PorterStemmer.MARTIN_EXTENSIONS)
-    stops = set(nltk.corpus.stopwords.words('english'))
-    to_strip = ' ()!@#$%^&*-+=|{}[]<>./?;:'
-
-    # match numbers (doesn't catch sci notation...)
-    numeric_regex = re.compile('(^\d+\.\d+$)|(^\d+$)')
-
-    d = defaultdict(set)
-    for sample, row in df.iterrows():
-        for value in row.values:
-            for part in value.split():
-                for w in part.split(','):
-                    tostem = w.strip(to_strip)
-                    if tostem in stops or '/' in tostem:
-                        # / is reserved as it's part of a URL
-                        continue
-                    if numeric_regex.match(tostem) is None:
-                        d[p.stem(tostem).lower()].add(sample)
-
-    return dict(d)
